// surrealist-surrealist-v3.9.7/src/types.tsx
export type ViewPage =
	| "dashboard"
	| "monitor"
	| "query"
	| "explorer"
	| "graphql"
	| "designer"
	| "authentication"
	| "functions"
	| "parameters"
	| "migrations"
	| "scheduler"
	| "booking"
	| "settings/general"
	| "settings/databases"
	| "settings/data"
	| "settings/capabilities"
	| "settings/version"
	| "settings/compute"
	| "settings/backups";

export type ConnectionSettingsTab =
	| "general"
	| "databases"
	| "data"
	| "capabilities"
	| "version"
	| "compute"
	| "backups";